Inputs: add "motion sensor" input type.
Motion sensor will switch on relay when it is turned on and switch off relay
when it turned off.
Relay with connected motion sensor input will not restore relay state
after reset, but it will align with input state.
Only "turn on" and "turn off" action triggers are available for motion
sensor input type.